+++ lib/etcd-discovery/service.rb
@@ -3,9 +3,9 @@ module EtcdDiscovery
   end
 
   class Service
-    attr_accessor :attributes
+    attr_accessor :attributes, :shard
 
-    def initialize(arg)
+    def initialize(arg, shard = nil)
       if arg.is_a? Etcd::Node
         @attributes = JSON.parse arg.value
       elsif arg.is_a? Hash
@@ -13,39 +13,55 @@ module EtcdDiscovery
       else
         raise TypeError, "requires a Etcd::Node or a Hash, not a #{arg.class}"
       end
+      @shard = shard
     end
 
-    def self.get(service)
+    def self.get(service, shard: nil)
       raise TypeError, "service should be a String, is a #{service.class}" unless service.is_a? String
 
       client = EtcdDiscovery.config.client
       begin
         service = client.get("/services_infos/#{service}")
-        new service.node
+        new(service.node, shard)
       rescue Etcd::KeyNotFound
-        new("name" => service)
+        new({"name" => service}, shard)
       end
     end
 
     def all
       client = EtcdDiscovery.config.client
+
       begin
         node = client.get("/services/#{attributes["name"]}", recursive: true).node
       rescue Etcd::KeyNotFound
         raise ServiceNotFound, attributes["name"]
       end
+
       raise ServiceNotFound, attributes["name"] if node.children.empty?
-      node.children.map { |c| Host.new(c) }
+
+      hosts = node.children.map { |c| Host.new(c) }
+      return hosts if @shard.nil?
+
+      hosts = hosts.select { |host| host.attributes["shard"].to_s == @shard.to_s }
+      raise ServiceNotFound, attributes["name"] if hosts.empty?
+
+      hosts
     end
 
     def one
       all.sample
     end
 
-    def to_uri(schemes = ["https", "http"])
+    def to_uri(schemes = ["https", "http"], **kwargs)
+      unless kwargs.empty?
+        raise TypeError, "to_uri does not accept keyword arguments; pass schemes as a String or an Array"
+      end
+
+      raise TypeError, "schemes should be a String or an Array" if schemes.is_a? Hash
+
       schemes = Array(schemes)
 
-      return one.to_uri(schemes) unless attributes["public"]
+      return one.to_uri(schemes) if @shard || !attributes["public"]
 
       scheme = schemes.find { |s| attributes["ports"][s] }
       raise "No valid scheme found" unless scheme