+++ spec/etcd_helper.rb
@@ -21,18 +21,29 @@ module EtcdHelper
       )
   end
 
-  def mock_hosts(service, value, count)
+  def mock_hosts(service, value, count = nil)
+    hosts = if value.is_a?(Array)
+      value
+    else
+      raise ArgumentError, "count is required when value is not an Array" if count.nil?
+
+      Array.new(count) do |i|
+        host = JSON.parse(value.to_json)
+        host["name"] = "#{service}-#{i}.scalingo.test"
+        host
+      end
+    end
+
     WebMock.stub_request(:get, "http://localhost:2379/v2/keys/services/#{service}?recursive=true").to_return do
       {
         body: {
           action: "get",
           node: {
             createdIndex: 1, modifiedIndex: 1, dir: true, key: "/services/#{service}",
-            nodes: Array.new(count) do |i|
-              value["name"] = "#{service}-#{i}.scalingo.test"
+            nodes: hosts.map do |host|
               {
                 createdIndex: 2, modifiedIndex: 2, dir: false,
-                value: value.to_json
+                value: host.to_json
               }
             end
           }